Finance

+++ Assets/Scripts/ObtenerDatos.cs
@@ -168,7 +168,7 @@ public class ObtenerDatos : MonoBehaviour
         userController.user_data["user_id"]=progreso.id_usuario;
         userController.user_data["capital"]=(int) progreso.dinero;
         userController.user_data["financiamiento"]=progreso.financiamiento;
-        userController.user_data["deuda"]=progreso.deuda;
+        userController.user_data["deuda"]=(int)progreso.deuda;
         mapManager.SetCycle(progreso.ciclo);
     }
